\name{data_origin1}
\alias{data_origin1}
\docType{data}
\title{An example set of ITC data as exported by Origin7 data sheet}
\description{
This is the content of the B110711CD4.txt file in the extdata folder, imported as a dataframe. 
}
\usage{data(data_origin1)}
\format{
  A data frame with observations on the following 7 variables.
  \describe{
    \item{\code{DH}}{a character vector}
    \item{\code{INJV}}{a character vector}
    \item{\code{Xt}}{a character vector}
    \item{\code{Mt}}{a character vector}
    \item{\code{XMt}}{a character vector}
    \item{\code{NDH}}{a character vector}
    \item{\code{B}}{a character vector}
  }
}

\examples{data(data_origin1)}
\keyword{datasets}
